+++ src/doomstat.c
@@ -184,7 +184,7 @@ int hudcolor_ms_comp;
 int event_timers[NUMTIMERS];
 
 // Enemies
-int extra_gibbing;
+int extra_gibbing_on;
 int bloodier_gibbing;
 int zdoom_item_drops;
 
@@ -205,6 +205,7 @@ int weapon_inertia_scale_pct;
 int sx_fix;
 int blink_keys;
 int automap_overlay_darkening;
+int extra_gibbing[NUMEXGIBS];
 
 int nugget_comp[NUGGET_COMP_TOTAL], default_nugget_comp[NUGGET_COMP_TOTAL];