+++ src/p_inter.c
@@ -703,18 +703,23 @@ static boolean P_NuggetExtraGibbing(mobj_t *source, mobj_t *target)
   extern fixed_t P_AproxDistance();
   extern void A_Punch(), A_Saw(), A_FireShotgun2();
 
-  if (casual_play && extra_gibbing && source && source->player
-      &&
-      (  (source->player->psprites->state->action.p2 == (actionf_p2)A_Punch
-          && source->player->powers[pw_strength]
-          && (P_AproxDistance(target->x - source->x, target->y - source->y)
-              < ((64*FRACUNIT) + target->info->radius)))
-       ||(source->player->psprites->state->action.p2 == (actionf_p2)A_Saw
-          && (P_AproxDistance(target->x - source->x, target->y - source->y)
-              < ((65*FRACUNIT) + target->info->radius)))
-       ||(source->player->psprites->state->action.p2 == (actionf_p2)A_FireShotgun2
-          && (P_AproxDistance(target->x - source->x, target->y - source->y)
-              < ((128*FRACUNIT) + target->info->radius)))
+  if (casual_play && extra_gibbing_on && source && source->player
+      && (
+          (extra_gibbing[EXGIB_FIST]
+           && source->player->psprites->state->action.p2 == (actionf_p2)A_Punch
+           && source->player->powers[pw_strength]
+           && (P_AproxDistance(target->x - source->x, target->y - source->y)
+               < ((64*FRACUNIT) + target->info->radius)))
+
+       || (extra_gibbing[EXGIB_CSAW]
+           && source->player->psprites->state->action.p2 == (actionf_p2)A_Saw
+           && (P_AproxDistance(target->x - source->x, target->y - source->y)
+               < ((65*FRACUNIT) + target->info->radius)))
+
+       || (extra_gibbing[EXGIB_SSG]
+           && source->player->psprites->state->action.p2 == (actionf_p2)A_FireShotgun2
+           && (P_AproxDistance(target->x - source->x, target->y - source->y)
+               < ((128*FRACUNIT) + target->info->radius)))
       )
      )
   { return true; }